requirements.txt
fastapi
uvicorn
python-dotenv
supabase
streamlit
pandas
openai
requests
